feat: nested tree folder structure
This feature is designed to support a folder/file structure that most
people are familiar with.  Instead of just having every single file in a
large list, we nest files inside of folders.

The main benefit here is that for large repos with a lot of files we
don't ask end-users to download a massive MB tree file.

It's more clicking -- which I don't like -- but this seems necessary
when there are 10k+ files in a project.